+++ libft/libft.h
@@ -39,7 +39,7 @@ typedef struct	s_fd
 
 int				get_next_line(const int fd, char **line);
 void			ft_bzero(void *s, size_t n);
-int				ft_atoi(const char *str);
+intmax_t		ft_atoi(const char *str);
 int				ft_isalnum(int c);
 int				ft_isalpha(int c);
 int				ft_isascii(int c);
